fix NPE

+++ hbase0.98/hbase-table-stat-0.98/src/main/java/com/kakao/hbase/stat/load/Load.java
@@ -145,7 +145,7 @@ public class Load {
     }
 
     public boolean isSummaryChanged(LoadEntry loadEntry) {
-        return summaryChangeMap.get(loadEntry).intValue() == ChangeState.changed.ordinal();
+        return summaryChangeMap.get(loadEntry) == ChangeState.changed.ordinal();
     }
 
     public boolean isDiffFromStart() {
